Extract `TeleconsultationMessageBuilder` interface
Rename existing `TeleconsultationMessageBuilder` to `LongTeleconsultMessageBuilder`
