+++ app/src/main/java/org/simple/clinic/summary/teleconsultation/messagebuilder/TeleconsultationMessageBuilder.kt
@@ -1,0 +1,7 @@
+package org.simple.clinic.summary.teleconsultation.messagebuilder
+
+import org.simple.clinic.summary.PatientTeleconsultationInfo
+
+interface TeleconsultationMessageBuilder {
+  fun message(patientTeleconsultationInfo: PatientTeleconsultationInfo): String
+}
